feat!: new lua API (using metatables)
This deprecates the old API and will be removed in the next tagged
releases probably v0.7

Currently every Lua API is a function call, which is fine, but most of
them are similar to others only having different arguments and uses same
call under the hood. This is not extensible if we want to create a new
mode #17 and have slight maintenance burden.

Using setmetatable we can probably build-up the API as we go. This will
be extensible and have less maintenance overhead.

This also changes the `ctype` utility. If you are using it then you
might wanna migrate to the following:

- `U.ctype.line` → `U.ctype.linewise`
- `U.ctype.block` → `U.ctype.blockwise`

Resolves #180